Fix code validation in Shipwar
Inserted explicit error throws to halt execution after certain conditions in the finish function. Also corrected the validation logic for online code input, adjusting the valid range for 'firstCode' from 1 to 4 instead of 0 to 3.
We're planning to rewrite all codes of Shipwar. It may be deployed before year 2100 or never.